integration-test-environment/test-orchestrator/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.3
requests==2.31.0
pytest==8.0.0
pytest-asyncio==0.23.5
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
prometheus-client==0.20.0
asyncio-timeout==4.0.3
aiofiles==23.2.1